pyproject.toml
[tool.poetry]
name = "finance-flow"
version = "0.1.0"
description = ""
authors = ["Nikolay Gendin <gendin786@gmail.com>"]
license = "MIT License"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.32.3"
python-dotenv = "^1.0.1"
pandas = "^2.2.3"
openpyxl = "^3.1.5"
python-dateutil = "^2.9.0.post0"


[tool.poetry.group.lint.dependencies]
flake8 = "^7.1.1"
mypy = "^1.13.0"
black = "^24.10.0"
isort = "^5.13.2"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
pytest-cov = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
disallow_untyped_defs = true
no_implicit_optional = true
warn_return_any = true
ignore_missing_imports = true
exclude = ".venv"

[tool.black]
# Максимальная длина строки
line-length = 119
exclude = '''
(
  /(
      \.git
  )/
)
'''

[tool.isort]
# максимальная длина строки
line_length = 119

[tool.pytest.ini_options]
pythonpath = ["src"]